move theme functions to utility
